+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Current development]
 * apply service migration for default alexa skill icons feature
 * update documentation of IURLSupplier
+* add new function relative_date
 ## [Releases]
 
 ## 0.22.6.0 - 2021-10-20